+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Issue #38:
   * `%with_inputs` and `%with_password` now wait 200 milliseconds each time (a constant), before 
     prompting user with an input in the Jupyter Notebook.
+  * Added `gonbui.RequestInput`, that will prompt the user with a text field in the notebook.
 
 ## 0.7.3 - 2023/07/14